+++ src/FakeRate.cpp
@@ -49,8 +49,52 @@ void FakeRate::toggle(CCNode* node, bool enabled) {
 }
 
 CCPoint FakeRate::getDIBOffset(int difficulty, GJDifficultyName name) {
-    if (difficulty < 1 || difficulty > 20) return { 0.0f, 0.0f };
-    return name == GJDifficultyName::Long ? LONG_OFFSETS[difficulty - 1] : SHORT_OFFSETS[difficulty - 1];
+    if (name == GJDifficultyName::Long) switch (difficulty) {
+        case 1: return { 0.0f, -5.0f };
+        case 2: return { 0.125f, -5.0f };
+        case 3: return { 0.0f, -5.0f };
+        case 4: return { 0.0f, -5.125f };
+        case 5: return { 0.25f, -5.0f };
+        case 6: return { 0.125f, -4.75f };
+        case 7: return { 0.0f, -5.0f };
+        case 8: return { 0.0f, -4.125f };
+        case 9: return { -0.125f, -4.125f };
+        case 10: return { 0.0f, -4.0f };
+        case 11: return { -0.125f, -4.125f };
+        case 12: return { 0.0f, -4.125f };
+        case 13: return { 0.125f, -4.125f };
+        case 14: return { 0.0f, -4.125f };
+        case 15: return { 0.0f, -4.125f };
+        case 16: return { 0.0f, -3.625f };
+        case 17: return { 0.0f, -3.625f };
+        case 18: return { 0.0f, -3.5f };
+        case 19: return { 0.0f, -3.5f };
+        case 20: return { 0.0f, -3.5f };
+    }
+    else if (name == GJDifficultyName::Short) switch (difficulty) {
+        case 1: return { -0.125f, -0.25f };
+        case 2: return { -0.125f, -0.25f };
+        case 3: return { -0.125f, -0.25f };
+        case 4: return { -0.125f, -0.375f };
+        case 5: return { -0.125f, -0.25f };
+        case 6: return { -0.125f, -0.25f };
+        case 7: return { -0.125f, -0.375f };
+        case 8: return { -0.125f, 0.5f };
+        case 9: return { -0.125f, 0.5f };
+        case 10: return { -0.125f, 0.25f };
+        case 11: return { -0.125f, 0.5f };
+        case 12: return { 0.125f, 0.5f };
+        case 13: return { 0.125f, 0.5f };
+        case 14: return { 0.125f, 0.5f };
+        case 15: return { 0.0f, 0.5f };
+        case 16: return { 0.0f, 1.25f };
+        case 17: return { 0.0f, 1.25f };
+        case 18: return { 0.0f, 1.125f };
+        case 19: return { 0.0f, 1.125f };
+        case 20: return { 0.0f, 1.125f };
+    }
+
+    return { 0.0f, 0.0f };
 }
 
 int FakeRate::getGRDOverride(CCSprite* sprite) {